requirements.txt
# Core Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Signal Processing
wfdb>=4.1.0
mne>=1.5.0

# Network Analysis
networkx>=3.1
python-louvain>=0.16
powerlaw>=1.5

# Machine Learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Statistical Analysis
statsmodels>=0.14.0
pingouin>=0.5.3

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0

# Optional: Deep Learning
# torch>=2.0.0
# tensorflow>=2.12.0